refactor(constants): 상수 파일 분리
- 상수가 많아서 파일 분리하는 것이 관리에 수월하다고 판단함
