+++ src/constants/amountConstants.js
@@ -1,0 +1,2 @@
+export const MIN_UNIT = 1_000;
+export const MAX_AMOUNT = 100_000;
