+++ src/constants/responseConstants.js
@@ -1,0 +1,4 @@
+export const YES = "y";
+export const NO = "n";
+export const UPPER_YES = YES.toUpperCase();
+export const UPPER_NO = NO.toUpperCase();
